+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/composable/common/button/sweep/GetEmulateProgressBrush.kt
@@ -1,5 +1,6 @@
 package com.flipperdevices.keyscreen.emulate.composable.common.button.sweep
 
+import androidx.compose.animation.core.Animatable
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.RepeatMode
 import androidx.compose.animation.core.animateFloat
@@ -7,15 +8,19 @@ import androidx.compose.animation.core.infiniteRepeatable
 import androidx.compose.animation.core.rememberInfiniteTransition
 import androidx.compose.animation.core.tween
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.SolidColor
+import com.flipperdevices.core.ktx.jre.toIntSafe
 import com.flipperdevices.keyscreen.emulate.model.EmulateProgress
 import kotlin.math.max
 
 private const val PROGRESS_BAR_END_PERCENT_FIXED_DELTA = 0.05f
 private const val PROGRESS_BAR_END_PERCENT_INFINITE = 0.2f
+private const val PROGRESS_BAR_END = 1.0f
 private const val PROGRESS_BAR_END_DURATION_MS = 1000
 
 @Composable
@@ -28,10 +33,15 @@ fun getEmulateProgressBrush(
         null -> {
             SolidColor(cursorColor)
         }
-        is EmulateProgress.Fixed -> fixedSizeBrush(
+        is EmulateProgress.Growing -> growingBrush(
             backgroundColor,
             cursorColor,
-            emulateProgress.toProgressFloat()
+            emulateProgress.duration
+        )
+        is EmulateProgress.GrowingAndStop -> growingBrushOneTime(
+            backgroundColor,
+            cursorColor,
+            emulateProgress.duration
         )
         EmulateProgress.Infinite -> animatedBrush(backgroundColor, cursorColor)
     }
@@ -39,11 +49,58 @@ fun getEmulateProgressBrush(
 
 private const val FIXED_SIZE_BRUSH_ANGEL = -90f
 
-private fun fixedSizeBrush(backgroundColor: Color, cursorColor: Color, progress: Float): Brush {
+@Composable
+private fun growingBrush(
+    backgroundColor: Color,
+    cursorColor: Color,
+    duration: Long
+): Brush {
+    val fixedProgress by rememberInfiniteTransition().animateFloat(
+        initialValue = 0f,
+        targetValue = PROGRESS_BAR_END,
+        animationSpec = infiniteRepeatable(
+            animation = tween(
+                durationMillis = duration.toIntSafe(),
+                easing = LinearEasing
+            ),
+            repeatMode = RepeatMode.Restart
+        )
+    )
+
     return rotatableSweepGradient(
         0f to cursorColor,
-        max(0f, progress - PROGRESS_BAR_END_PERCENT_FIXED_DELTA) to cursorColor,
-        progress to backgroundColor,
+        max(0f, fixedProgress - PROGRESS_BAR_END_PERCENT_FIXED_DELTA) to cursorColor,
+        fixedProgress to backgroundColor,
+        angel = FIXED_SIZE_BRUSH_ANGEL
+    )
+}
+
+@Composable
+private fun growingBrushOneTime(
+    backgroundColor: Color,
+    cursorColor: Color,
+    duration: Long
+): Brush {
+    val fixedProgressState = remember { Animatable(0f) }
+    LaunchedEffect(Unit) {
+        fixedProgressState.animateTo(
+            targetValue = PROGRESS_BAR_END,
+            animationSpec = tween(
+                durationMillis = duration.toIntSafe(),
+                easing = LinearEasing
+            )
+        )
+    }
+    val fixedProgress = fixedProgressState.value
+
+    if (fixedProgress == PROGRESS_BAR_END) {
+        return animatedBrush(backgroundColor, cursorColor)
+    }
+
+    return rotatableSweepGradient(
+        0f to cursorColor,
+        max(0f, fixedProgress - PROGRESS_BAR_END_PERCENT_FIXED_DELTA) to cursorColor,
+        fixedProgress to backgroundColor,
         angel = FIXED_SIZE_BRUSH_ANGEL
     )
 }