+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/model/EmulateProgress.kt
@@ -8,9 +8,8 @@ sealed class EmulateProgress {
     object Infinite : EmulateProgress()
 
     @Stable
-    data class Fixed(val progress: Long, val total: Long) : EmulateProgress() {
-        fun toProgressFloat() = if (total != 0L) {
-            progress.toFloat() / total.toFloat()
-        } else 0f
-    }
+    data class Growing(val duration: Long) : EmulateProgress()
+
+    @Stable
+    data class GrowingAndStop(val duration: Long) : EmulateProgress()
 }