+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/tasks/CloseEmulateAppTask.kt
@@ -5,6 +5,8 @@ import com.flipperdevices.bridge.service.api.provider.FlipperServiceProvider
 import com.flipperdevices.core.ui.lifecycle.OneTimeExecutionBleTask
 import com.flipperdevices.keyscreen.emulate.viewmodel.helpers.EmulateHelper
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.filter
+import kotlinx.coroutines.flow.first
 
 class CloseEmulateAppTask(
     serviceProvider: FlipperServiceProvider,
@@ -18,7 +20,9 @@ class CloseEmulateAppTask(
         input: Unit,
         stateListener: suspend (Unit) -> Unit
     ) {
-        emulateHelper.stopEmulate(serviceApi.requestApi)
+        emulateHelper.stopEmulateForce(serviceApi.requestApi)
+        // Waiting until the emulation stops
+        emulateHelper.getRunningState().filter { !it }.first()
         stateListener(Unit)
     }